css/css-values/clamp-opacity-down.html

<!DOCTYPE html>
<meta charset=utf-8>
<html>
<head>
<title>CSS Values and Units: Range checking clamped values</title>
<link rel="author" title="Alan Stearns" href="mailto:stearns@adobe.com">
<link rel="help" href="https://drafts.csswg.org/css-values/#calc-range">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
	.test { border: thick solid; opacity: calc(-10); animation: bounce 2s infinite linear;}
	@keyframes bounce {
	  0% { opacity: calc(1); }
	}
</style>
</head>
<body>
	<div id='testDiv'>foo</div>
</body>
</html>
<script>
	function checkValue() {
		return (window.getComputedStyle(testDiv)["opacity"]);
	}

	async_test(function(t) {
		testDiv.setAttribute("class", "test");
		setTimeout('clampedValue = checkValue();', 1500);
		t.step_timeout(function() { 
			assert_equals ( clampedValue, "0" );
			t.done();
		}, 2000);
	})
</script>
<!DOCTYPE html>
